+++ predict_folder.py
@@ -30,10 +30,10 @@ if __name__ == '__main__':
         checkpoint = torch.load(path, map_location="cpu")
         state_dict = checkpoint.get("state_dict", checkpoint)
         model.load_state_dict(
-            {re.sub("^module.", "", k): v for k, v in state_dict.items()}, strict=True)
+            {re.sub("^module.", "", k): v.to(torch.float32) for k, v in state_dict.items()}, strict=True)
         model.eval()
         del checkpoint
-        models.append(model.half())
+        models.append(model)
 
     frames_per_video = 32
     video_reader = VideoReader()
